filtered search option for map points
with this endpoint the backend gets an JSON object from the frontend with the parameters that the user wants to search for, then a GET call to the database is made with those parameters and null values for the ones not included in the search. The search is made with an AND operator between parameters.